Fix scrolling to a local fragment
Problem: on navigation to a local fragment, the web-page is scrolled, and the fixed web-page header overlaps the title of the fragment.
Solution: use the css property 'scroll-padding-top' in 'html' element (that is the main document scroller) to reserve space for the header.  Reserve additional space when an element 'a' is the target, since it's placed after the actual fragment title. Remove the old solution (see commit#42c301bc) that doesn't work.
Bounce: decorate the target element with bottom dotted line to make it more visible.

Fixes Issue#13