Fix scrolling to a local fragment, tune
Due to a mistake in a LESS code fragment, a generated CSS selector is: `a[name] :target` (i.e. a descendant of `a[name]` that is a target), but we need `a[name]:target` (i.e. `a[name]` that is a target).

Also, don't apply dotted bottom border to a target element when it is a `div` or `section` element, since usually they envelope both the title and body of a fragment. But we want to decorate only a title (if any).